Use state_values_v2 for Android LTV pipeline (#4887)
* Don't try to write existing view files

* Use state_values_v2 for client ad click predictions

* Normalize countries in client_ltv

* Don't get view if unavailable

* Add test for new version of existing table

* Fully qualify tables in view defn

+++ bigquery_etl/cli/query.py
@@ -170,12 +170,14 @@ def create(ctx, name, sql_dir, project_id, owner, init, dag, no_schedule):
     path = Path(sql_dir)
 
     if dataset.endswith("_derived"):
-        # create a directory for the corresponding view
+        # create directory for this table
         derived_path = path / project_id / dataset / (name + version)
         derived_path.mkdir(parents=True)
 
+        # create a directory for the corresponding view
         view_path = path / project_id / dataset.replace("_derived", "") / name
-        view_path.mkdir(parents=True)
+        # new versions of existing tables may already have a view
+        view_path.mkdir(parents=True, exist_ok=True)
     else:
         # check if there is a corresponding derived dataset
         if (path / project_id / (dataset + "_derived")).exists():
@@ -193,9 +195,9 @@ def create(ctx, name, sql_dir, project_id, owner, init, dag, no_schedule):
 
     click.echo(f"Created query in {derived_path}")
 
-    if view_path:
+    if view_path and not (view_file := view_path / "view.sql").exists():
+        # Don't overwrite the view_file if it already exists
         click.echo(f"Created corresponding view in {view_path}")
-        view_file = view_path / "view.sql"
         view_dataset = dataset.replace("_derived", "")
         view_file.write_text(
             reformat(
